Add files via upload
In this task, we had to use the MINST database to test the accuracy of a simple neural network. This would require the database to be split into a training and validation dataset and then to create a model to be trained on the database. The model had to achieve a minimal accuracy of 97%.