Add files via upload
This task was more complex than the rest, as it required me to use a YOLO network to detect the drones in the images and draw bounding boxes around them. Then it required for the model to be evaluated. This was more complex than the other tasks because it required me to clone the YOLOv5 repository and change certain lines in specific files to make it work for drones, as well as changing the database it was running on.